工程结构与u2稳定性改善
