+++ basic_uia/device.py
@@ -2,7 +2,6 @@ import os
 
 import config as cf
 from basic_uia import logger
-# from .uiautomator import Device
 import uiautomator2 as u2
 
 TEST_DEVICE = None
@@ -17,25 +16,18 @@ def confirm_device_connection(device_id):
         raise ConnectionError('device {} not found.'.format(device_id))
 
 
-class DeviceItem(object):
-    """ 设备类 """
-    def __init__(self, device_id):
-        self.device_id = device_id
-        confirm_device_connection(device_id)
-        self._device_instance = u2.connect(device_id)
-        logger.info(self._device_instance.device_info)
-        logger.info('{} connected'.format(device_id))
-
-    def __getattr__(self, item):
-        return getattr(self._device_instance, item)
-
-    def __call__(self, *args, **kwargs):
-        return self._device_instance.__call__(*args, **kwargs)
+def get_device(device_id):
+    """ 获取设备对象 """
+    confirm_device_connection(device_id)
+    device_instance = u2.connect(device_id)
+    logger.info(device_instance.device_info)
+    logger.info('{} connected'.format(device_id))
+    return device_instance
 
 
 def connect_device():
     """ 关联所有配置的设备 """
     device_id = cf.DEVICE_ID_LIST[0]
     global TEST_DEVICE
-    TEST_DEVICE = DeviceItem(device_id)
+    TEST_DEVICE = get_device(device_id)
     logger.info('Device [ {} ] connected.'.format(device_id))
