+++ basic_uia/logger.py
@@ -26,16 +26,3 @@ def init_logger(logger_path):
     info = logging.info
     error = logging.error
     info(' SYSTEM READY '.center(cf.LENGTH_OF_SPLIT_LINE, '='))
-
-
-def add_log(func):
-    """ 函数执行流程的日志装饰器 """
-    @functools.wraps(func)
-    def deco(*args, **kwargs):
-        func_desc = func.__doc__
-        logging.info('*** Start Function [{}] ***'.format(func_desc))
-        start_time = time.time()
-        result = func(*args, **kwargs)
-        logging.info('--- Done in {} seconds ---'.format(str(round(time.time() - start_time, 2))))
-        return result
-    return deco
